UI Changes

+++ app/src/main/java/uart/terminal/androidstudio/com/myapplication/Values.java
@@ -4,5 +4,5 @@ public class Values {
     public static int baudrate;
     public static int dataBits;
     public static String parity;
-    public static int stopBits;
+    public static double stopBits;
 }
